Completed the encoding of whole secret file size

+++ encode.h
@@ -70,7 +70,7 @@ Status encode_magic_string(const char *magic_string, EncodeInfo *encInfo);
 Status encode_secret_file_extn(const char *file_extn, EncodeInfo *encInfo);
 
 /* Encode secret file size */
-Status encode_secret_file_size(long file_size, EncodeInfo *encInfo);
+Status encode_secret_file_size(int file_size, EncodeInfo *encInfo);
 
 /* Encode secret file data*/
 Status encode_secret_file_data(EncodeInfo *encInfo);
